% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Go2Web-package.R
\docType{package}
\name{Go2Web-package}
\alias{Go2Web-package}
\title{Go2Web-package}
\description{
Go2Web-package
}
\details{
This package publish the content of the current directory on a web server created with python.
Technically, it runs the system command "python -m SimpleHTTPServer 8000 &>/dev/null".
The default port is 8000 and every output is redirected to /dev/null (i.e. not shown to the user).
Both options can be changed by the user.
}

