% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{Go2Web}
\alias{Go2Web}
\title{Go2Web}
\usage{
Go2Web(port = 8000, redirect.to.null = TRUE)
}
\arguments{
\item{port}{Defaults to 8000}

\item{redirect.to.null}{Defaults to TRUE}
}
\description{
This function publish the content of the current directory on a web server. Technically,
it runs the system command "python -m SimpleHTTPServer 8000 &>/dev/null". The default port is 8000 and
every output is redirected to /dev/null (i.e. not shown to the user). Both options can be changed by the user.
}
\examples{
Go2Web()
}

