% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{webserver.status}
\alias{webserver.status}
\title{webserver.status}
\usage{
webserver.status()
}
\description{
This function checks if a python SimpleHTTPServer is running.
}
\examples{
webserver.status()
}

