% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/functions.R
\name{webserver.stop}
\alias{webserver.stop}
\title{webserver.stop}
\usage{
webserver.stop(port = "8000")
}
\arguments{
\item{port}{Defaults to 8000}
}
\description{
This function stops the web server started with ``webserver.start()''. The default port is 8000 and can be changed by the user to whatever port the server is running on.
}
\examples{
webserver.stop()
}

