\name{Go2Web-package}
\alias{Go2Web-package}
\alias{Go2Web}
\docType{package}
\title{
\packageTitle{Go2Web}
}
\description{
\packageDescription{Go2Web}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Go2Web}
\packageIndices{Go2Web}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Go2Web}

Maintainer: \packageMaintainer{Go2Web}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
